@@ -3,31 +3,28 @@
   "version": "0.1.0",
   "homepage": "https://sqren.github.io/countdown.run/",
   "private": true,
-  "eslintConfig": {
-    "extends": ["react-app", "standard"]
-  },
   "dependencies": {
     "classnames": "^2.2.5",
     "lodash.debounce": "^4.0.8",
     "lodash.omit": "^4.5.0",
     "lodash.once": "^4.1.1",
     "lodash.padstart": "^4.6.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-redux": "^5.0.3",
-    "redux": "^3.6.0"
+    "prop-types": "^15.6.1",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "react-redux": "^5.0.7",
+    "redux": "^4.0.0"
   },
   "devDependencies": {
-    "eslint": "^3.18.0",
-    "eslint-config-standard": "^7.1.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-node": "^4.2.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^2.1.1",
-    "gh-pages": "^0.12.0",
-    "node-sass": "^4.5.1",
-    "npm-run-all": "^4.0.2",
-    "react-scripts": "0.9.5"
+    "eslint": "^4.19.1",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.7.0",
+    "gh-pages": "^1.1.0",
+    "prettier": "^1.12.1",
+    "node-sass": "^4.8.3",
+    "npm-run-all": "^4.1.2",
+    "react-scripts": "1.1.4"
   },
   "scripts": {
     "predeploy": "npm run build",
