@@ -1,4 +1,4 @@
-export default function initServiceWorker (store) {
+export default function initServiceWorker () {
   if ('serviceWorker' in navigator) {
     window.addEventListener('load', () => {
       navigator.serviceWorker.register('/sw.js')
